refactor: split lib.rs into focused modules
- Extract types into types.rs (AppEvent, NetworkCommand, Message, etc.)
- Move configuration to config.rs (relay URLs, defaults)
- Create utils.rs for utility functions
- Organize commands in commands.rs (input processing, command handling)
- Extract group management to groups.rs (creation, invites, welcomes)
- Move message handling to messages.rs (sending, processing, fetching)
- Create network.rs for network operations (key packages)
- Extract profile management to profiles.rs (display names, fetching)
- Move state/initialization logic to state.rs

The main lib.rs now only contains the struct definition and constructor,
making the codebase more maintainable and organized (90% confidence).
